rpms: New function check_file_modification()
This function is used to invoke rpm to check if the given file was
modified in the source operating system and if we need to update it or
the rpm will take care of its update.

This was initially implemented as part of the OpenSshConfigScanner
actor, but it is usable also in other actors.
